Modify Makefile to support flags from env
The Makefile of the project does not support flags like
CFLAGS from the env. This is a problem since the debhelper
tools set their flags in a env variable and not as command
line flags for make.

This patch modifies the Makefiles to use VAR+=... instead of
VAR=... so the env is used, too.

Author: Jan Luca Naumann <j.naumann@fu-berlin.de>
Bug: https://github.com/Drive-Trust-Alliance/sedutil/issues/60
Forwarded: no
Last-Update: 2016-06-04